Created component `ki-glossary-entry`
New component `ki-glossary-entry` enhaces the presentation of the
glossary index list items by providing more relevant information.

It displays the name of the parent `ki-topic` or `ki-section` if available.

It also sets the whole lineage of the sections to topics as its
HTMLElement.title property, which would cause a tooltip containing the
information to be displayed on hover.